Authorization

+++ app/Http/Controllers/Backend/RoleController.php
@@ -7,6 +7,7 @@ use App\Models\Role;
 use App\Models\Module;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -17,6 +18,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.role.view');
+
         $roles = Role::all();
         return view('backend.roles.index', compact('roles'));
     }
@@ -28,6 +31,8 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('app.role.create');
+
         $modules = Module::all();
         return view('backend.roles.form', compact('modules'));
     }
@@ -40,6 +45,8 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('app.role.create');
+
         $this->validate($request, [
             'name' => 'required|unique:roles',
             'permissions' =>'required|array',
@@ -73,6 +80,8 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize('app.role.edit');
+
         $modules = Module::all();
         return view('backend.roles.form', compact('modules', 'role'));
     }
@@ -86,12 +95,14 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        Gate::authorize('app.role.edit');
+
         $role->update([
             'name' => $request->name,
             'slug' => Str::slug($request->name)
         ]);
 
-        $role->permissions()->sync($request->input('permissions'), []);
+        $role->permissions()->sync($request->input('permissions'));
         notify()->success("Role Updated","Success");
         return redirect()->route('app.roles.index');
     }
@@ -104,6 +115,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('app.role.delete');
+
         if($role->deletable){
             $role->delete();
             notify()->success("Role Deleted","Success");
